prevent nvidia hang by using ints, not floats for numProbes

+++ Engine/source/shaderGen/GLSL/shaderFeatureGLSL.cpp
@@ -2905,7 +2905,7 @@ void ReflectionProbeFeatGLSL::processPix(Vector<ShaderComponent*>& componentList
    //Reflection Probe WIP
    U32 MAX_FORWARD_PROBES = 4;
 
-   Var * numProbes = new Var("numProbes", "float");
+   Var * numProbes = new Var("numProbes", "int");
    numProbes->uniform = true;
    numProbes->constSortPos = cspPotentialPrimitive;
